Split out _cache.py and _util.py files
